requirements.txt
aiohttp
docopt
lxml
PyICU
pymongo
pypandoc
python-Levenshtein
PyYAML
